+++ collector-bitmovin-player/src/androidTest/java/com/bitmovin/analytics/bitmovin/player/Phone2PlayerParallelScenariosTest.kt
@@ -59,12 +59,10 @@ class Phone2PlayerParallelScenariosTest {
             val player1AnalyticsConfig =
                 TestConfig.createAnalyticsConfig(
                     backendUrl = mockedIngressUrl,
-                    ssaiEngagementTrackingEnabled = false,
                 )
             val player2AnalyticsConfig =
                 TestConfig.createAnalyticsConfig(
                     backendUrl = mockedIngressUrl,
-                    ssaiEngagementTrackingEnabled = false,
                 )
 
             val player1Config =