+++ collector/src/main/java/com/bitmovin/analytics/adapters/DefaultPlayerAdapter.kt
@@ -49,7 +49,7 @@ abstract class DefaultPlayerAdapter(
 
     // TODO [AN-4317]: this wiring is not good, we should aim for getting rid of the PlayerAdapter dependency
     private val ssaiEngagementMetricsService: SsaiEngagementMetricsService =
-        SsaiEngagementMetricsService(analytics = bitmovinAnalytics, analyticsConfig = config, playerAdapter = this, Handler(looper))
+        SsaiEngagementMetricsService(analytics = bitmovinAnalytics, playerAdapter = this, ssaiTimeoutHandler = Handler(looper))
     final override val ssaiService = SsaiService(stateMachine, ssaiEngagementMetricsService)
 
     init {